requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==1.11
django-celery==3.3.0
django-celery-beat==1.5.0
django-celery-results==1.1.1
django-chunked-upload==1.1.3
django-crispy-forms==1.7.2
django-formtools==2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
django-timezone-field==3.0
djangorestframework==3.8.2
ffmpeg-python==0.1.18
ffmpy==0.2.2
Flask==1.0.3
Flask-SocketIO==4.0.0
future==0.17.1
httpie==1.0.2
httplib2==0.9.2
idna==2.8
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
kombu==3.0.37
Markdown==3.1.1
MarkupSafe==1.1.1
meld3==1.0.2
oauthlib==3.0.1
Pillow==5.3.0
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.0
PySocks==1.6.8
python-crontab==2.3.6
python-dateutil==2.8.0
python-engineio==3.8.0
python-memcached==1.59
python-socketio==4.1.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sorl-thumbnail==12.5.0
supervisor==4.0.3
svgwrite==1.2.1
Tree==0.2.4
uritemplate==3.0.0
urllib3==1.24.3
vine==1.3.0
Werkzeug==0.15.4
